Restore AUTOMATION_ROUTE
