+++ crosscompute/macros/__init__.py
@@ -1,5 +1,6 @@
 # flake8: noqa
 from .disk import make_folder
+from .iterable import find_dictionary
 from .log import format_path
 from .process import StoppableProcess
 from .text import compact_whitespace, normalize_key
